+++ src/components/pages/HomePage/index.ts
@@ -1,0 +1,3 @@
+export * from './HomePage';
+export * from './DashboardCard';
+export * from './DashboardListView';