+++ src/components/pages/index.ts
@@ -1,1 +1,2 @@
 export * from './TVDashboard/TVDashboard';
+export * from './HomePage';