+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from './common/DataSource';
 export * from './common/TVRoute';
 export * from './common/WidgetConfig';
 export * from './api/ApiResponse';
+export * from './jira';