+++ src/types/jira/index.ts
@@ -1,0 +1,2 @@
+export * from './JiraIssue';
+export * from './JiraSearch';